edited css and html for profile.

+++ food/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from .models import *
+from django.contrib.auth.models import User
 from django.core.paginator import Paginator
 from django.utils import timezone 
 from .forms import *
@@ -9,29 +10,40 @@ from django.contrib.auth.decorators import login_required
 from accounts.models import *
 from allauth.socialaccount.models import SocialAccount, SocialToken
 import requests, json
+from django.db.models.functions import Replace
+from django.db.models import Value
 # Create your views here.
 
+def person(request):
+    profile = Profile.objects.get_username()
+    users = User.objects.get(user=request.user)
+    return render(request, 'base.html', {'profile':profile, 'users':users})
 
 def home(request):
     foods = Food.objects
     food_list = Food.objects.all()
-    print(type(food_list))
     paginator = Paginator(food_list,4)
     page = request.GET.get('page')
     posts = paginator.get_page(page)
     return render(request, 'home.html', {'foods':foods, 'posts':posts})
 
+# 고객이 보는 나의 요청내역
 def myCart(request):
-    requests = Cart.objects
     request_list = Cart.objects.all().filter(sender=request.user)
     # print(type(request_list))
     return render(request, 'myCart.html', {'request_list':request_list})
 
-# def myCart_for_owner(request):
-#   notification = Cart.objects
-#   notifications = Cart.objects.all().filter(reciever=request.user)
-#   return render(request, 'myCart.html, {'notifications':notifications})
-#   페이지를 따로 만들어야 하는지, 아님 같은 페이지안에 if문 줘서 local 회원이랑 social회원이랑 구분해서 안보이게 할 수 있는지 알아보기
+# 매장이 보는 나의 요청온내역 
+def requested_cart(request):
+    the_receiver = Profile.objects.get(user=request.user)
+    requested_list = Cart2.objects.all().filter(receiver=the_receiver)
+    return render(request, 'requested_cart.html', {'request_list':requested_list})
+
+# 매장이 보는 게시글관리 
+def post_retrieve(request):
+    the_author = Profile.objects.get(user=request.user)
+    post_list = Food.objects.all().filter(author=the_author)
+    return render(request, 'myPostList.html', {'post_list':post_list})
 
 def detail(request, food_id):
     food_detail = get_object_or_404(Food, pk=food_id)
@@ -41,22 +53,21 @@ def detail(request, food_id):
 def add_comment_to_food(request, food_id):
     food_detail = get_object_or_404(Food, pk=food_id)
     if request.method == 'POST':
-        form = CommentForm(request.POST, instance=food_detail)
+        form = CommentForm(request.POST)
         if form.is_valid():
             comment = form.save(commit=False)
-            comment.author = User.objects.get(username=request.user)
-            # 대체 왜... Comment 모델의 author는 Profile로 명시하지 않은 앤데.. 왜 Profile instance만 가능할까...
+            comment.author = request.user
             comment.post = food_detail
             comment.save()
             return redirect('detail', food_id=food_id)
         else:
-            form = CommentForm(instance=food_detail)
+            form = CommentForm()
         return render(request, 'comment.html', {'form':form})
 
+# 결제 전 고객이 자신의 요청 내역을 보는 카트를 저장
 @login_required
 def cart(request, food_id):
     food_detail = get_object_or_404(Food, pk=food_id)
-    # 현재 페이지의 author를 어떻게 가져오지?
     if request.method == 'POST':
         people = request.POST['people']
         total_price = request.POST['total_price']
@@ -70,16 +81,63 @@ def cart(request, food_id):
         # my내역에서 확인할 수 있도록 페이지 만들고 바꾸기
     return render(request, 'detail.html', {'error':'요청실패', 'food':food_detail})
 
+
+# 결제준비를 마친 고객이 결제 완료되었다는 문구를 보는 페이지 
+def successPage(request):
+    access_token = SocialToken.objects.get(account__user=request.user, account__provider="kakao").token
+    pg_token = request.GET['pg_token']
+    print("@@@@@@@@@@@@@@@@@@@@@@@@ pg_token : {} , ".format(pg_token))
+    cart2 = Cart2.objects.get(sender=request.user)
+    if request.method == 'POST':
+        return render(request, 'ordering.html')
+    URL = 'https://kapi.kakao.com/v1/payment/approve'
+    headers = {
+        'Authorization': 'Bearer ' + str(access_token),
+        'Content-Type':'application/x-www-form-urlencoded;charset=utf-8'
+    }
+    data = {
+        "cid": "TC0ONETIME",
+        "partner_order_id": cart2.order_id,
+        "partner_user_id": cart2.receiver,
+        "pg_token": pg_token,
+        "tid": cart2.tid
+    }
+    res = requests.post(URL,headers=headers,data=data)
+    print("@@@@@@ {} @@@@@@@@@@@ {}".format(res.status_code, cart2.status))
+    if res.status_code == 200:
+        cart2 = Cart2.objects.get(sender=request.user)
+        cart2.status = "1"
+        cart2.save()
+        return render(request, 'check_success.html')
+    else:
+        cart2 = Cart2.objects.get(sender=request.user)
+        cart2.status = "-1"
+        cart2.save()
+        return redirect('fail')
+
+# 결제준비가 된 고객의 주문을 매장이 볼 수 있도록 cart2모델에 저장하기 
+def successCart(request, cart_id):
+    cart_detail = get_object_or_404(Cart, pk=cart_id)
+    if request.method == 'POST':
+        sender = cart_detail.sender
+        receiver = cart_detail.receiver
+        people = cart_detail.people
+        total_price = cart_detail.total_price
+        request_date = timezone.now()
+        title = cart_detail.title
+        cart2 = Cart2(sender=sender, receiver=receiver, people=people, total_price=total_price, request_date=request_date, title=title)
+        cart2.save()
+        return redirect('home')
+    else:
+        return render(request, 'ordering.html', {'cart':cart_detail})
+
 def foodpost(request):
     if request.method == 'POST':
         form = FoodPost(request.POST, request.FILES)
         if form.is_valid():
             post = form.save(commit=False)
-            # post.author = request.user
-            # post.image = request.FILES['image']
             the_author = Profile.objects.get(user=request.user)
             post.author = the_author
-            # post.author = request.user
             post.pub_date=timezone.now()
             post.save()
             return redirect('home')
@@ -93,7 +151,6 @@ def edit(request, food_id):
         form = FoodPost(request.POST, request.FILES, instance=food_detail)
         if form.is_valid():
             post = form.save(commit=False)
-            # post.image = request.FILES['image']
             the_author = Profile.objects.get(user=request.user)
             post.author = the_author
             post.pub_date=timezone.now()
@@ -112,18 +169,20 @@ def search(request):
 
 def delete(request, food_id):
     deleted_food = get_object_or_404(Food, pk=food_id)
-    deleted_food.delete()
-    return redirect('home')
-    # 게시글 관리에서 삭제할 수 있게 옮기기
+    if request.method == 'POST':
+        deleted_food.delete()
+        return redirect('myposts')
 
+# 결제 준비 
 @login_required
 def checkplz(request):
     access_token = SocialToken.objects.get(account__user=request.user, account__provider="kakao").token
+    print(access_token)
     partner_user_id = Profile.objects.get(nickname=request.POST['receiver'])
     quantity = request.POST['people']
     total_amount = request.POST['total_price']
-    item_name = request.POST['title']
-    partner_order_id = request.POST['cart_id']
+    item_name = Food.objects.get(title=request.POST['title'])
+    partner_order_id = Cart.objects.get(id=request.POST['cart_id']).id
     data = {
         "cid": "TC0ONETIME",
         "partner_order_id": partner_order_id,
@@ -145,30 +204,94 @@ def checkplz(request):
     }
     res = requests.post(URL,headers=headers,data=data)
     ## 받아온 데이터를 Json => Dict으로 바꿈.(파이썬에서 사용 가능하게)
-    print(res.text)
-    resp = json.loads(res.text)
-    return redirect(resp['next_redirect_pc_url'])
+    if res.status_code == 200:
+        resp = json.loads(res.text)
+        cart2 = Cart2(
+            tid=resp['tid'],order_id=partner_order_id,
+            sender=request.user, receiver=partner_user_id, people=quantity, 
+            total_price=total_amount, request_date=timezone.now(), title=item_name,status="0")
+        cart2.save()
+        print("@@@@@@@@@@@ order id = {} , tid = {}".format(partner_order_id , resp['tid']))
+        return redirect(resp['next_redirect_pc_url'])
+    else:
+        return redirect('fail')
     
 
-# def success(request):
-#     return render(request, 'check_success.html')
+def checkCanceled(request, cart2_id):
+    cart2 = get_object_or_404(Cart2, order_id=cart2_id)
+    access_token = SocialToken.objects.get(account__user=cart2.sender, account__provider="kakao").token
+    if request.method == 'POST':
+        if request.POST['deny']:
+            cart2.status == "-1"
+            headers = {
+                    'Authorization': 'Bearer ' + str(access_token),
+                    'Content-Type':'application/x-www-form-urlencoded;charset=utf-8'
+                }
+            data = {
+                    "cid": "TC0ONETIME",
+                    "partner_order_id": cart2.order_id,
+                    "cancel_amount": cart2.total_price,
+                    "cancel_vat_amount": 0,
+                    "cancel_tax_free_amount": 0,
+                    "tid": cart2.tid
+                }
+            URL = 'https://kapi.kakao.com/v1/payment/cancel'
+            res = requests.post(URL,headers=headers,data=data)
+            cart2.save()
+            cart_of_sender = Cart.objects.all().filter(id=cart2_id)
+            cart_of_sender.delete()
+            return redirect('fail')
 
+# def reservation(request, cart2_id):
+#     cart2 = get_object_or_404(Cart2, order_id=cart2_id)
+#     access_token = SocialToken.objects.get(account__user=cart2.sender, account__provider="kakao").token
+#     if request.POST['ok']:
+#         headers = {
+#                     'Authorization': 'Bearer ' + str(access_token)
+#                 }
+#         receiver_uuids = {
+#             "diddnjs02"
+#         }
+#         title = Food.objects.get(title=cart2.title)
+#         image = title.image
+#         resAddress = title.author
+#         the_address = resAddress.address
+#         template_object = { 
+#                 "object_type": "location",
+#                 "content": {
+#                     "title": "카카오 판교오피스",
+#                     "description": "카카오 판교오피스 위치입니다.",
+#                     "image_url": "https://mud-kage.kakao.com/dn/drTdbB/bWYf06POFPf/owUHIt7K7NoGD0hrzFLeW0/kakaolink40_original.png",
+#                     "image_width": 800,
+#                     "image_height": 800,
+#                     "address": "경기 성남시 분당구 판교역로 235 에이치스퀘어 N동 7층",
+#                     "address_title": "카카오 판교오피스",
+#                 "link": {
+#                 "web_url": "http://dev.kakao.com",
+#                 "mobile_web_url": "http://dev.kakao.com/mobile",
+#                 "android_execution_params": "platform=android",
+#                 "ios_execution_params": "platform=ios"
+#                     }
+#                 },
+#                     {
+#                 "buttons": [
+#                         "title": "웹으로 보기",
+#                         "link": {
+#                             "web_url": "http://dev.kakao.com",
+#                             "mobile_web_url": "http://dev.kakao.com/mobile"
+#                             }
+#                     }
+#                 ],
+#                     }
+        
+#             URL = "https://kapi.kakao.com/v1/api/talk/friends/message/default/send"
+#             res = requests.post(URL, headers=headers, receiver_uuids=receiver_uuids, template_object=template_object)
+#             cart_of_sender = Cart.objects.all().filter(id=cart2_id)
+#             cart_of_sender.delete()
+#             return redirect('success')
+# 
+# 
 def fail(request):
     return render(request, 'check_fail.html')
 
-def successCart(request):
-    request_list = Cart.objects.get(sender=request.user)
-    if request.method == 'POST':
-        sender = User.objects.get(username=request.POST['sender'])
-        receiver = Profile.objects.get(nickname=request.POST['receiver'])
-        the_title = Food.objects.get(title=request.POST['title'])
-        people = request.POST['people']
-        total_price = request.POST['total_price']
-        # request_date = request.timezone.now()
-        successCart = SuccessCart(sender=sender, people=people, total_price=total_price, receiver=receiver, title=the_title)
-        successCart.save()
-        return redirect('home')
-    else:
-        return render(request, 'check_success.html', {'error':'매장에 요청 들어가기 실패', 'requests':request_list})
 
-
